Added explicit typecast in GPS tags to avoid compiler warnings.

+++ src/exif.cpp
@@ -312,7 +312,8 @@ void Exif::setExifEntry(ExifData *data, ExifTypedTag tag, const QVariant &value)
 
     case EXIF_FORMAT_RATIONAL:
 
-        switch (entry->tag) {
+        // GPS tags are not members of the ExifTag enum so we need the int cast
+        switch ((int)entry->tag) {
         case EXIF_TAG_GPS_LATITUDE:
         case EXIF_TAG_GPS_LONGITUDE: {
             ExifRational rat;
@@ -381,7 +382,8 @@ void Exif::setExifEntry(ExifData *data, ExifTypedTag tag, const QVariant &value)
 
 void Exif::updateReferenceTag(ExifTag tag, bool positive)
 {
-    switch (tag) {
+    // GPS tags are not members of the ExifTag enum so we need the int cast
+    switch ((int)tag) {
         case EXIF_TAG_GPS_LATITUDE:
             setEntry(QuillMetadata::Tag_GPSLatitudeRef, QVariant(positive ? "N" : "S"));
             break;
